Allow grabber to be grabbed at fractions

+++ include/maritime/servos.h
@@ -3,7 +3,7 @@
 
 void init_servos();
 void drop(int idx, int value);
-void grab(int value);
+void grab(float value);
 void shoot(int idx, int value);
 
 #endif /* SERVO_H */
